+++ app/health-landing/page.tsx
@@ -6,7 +6,7 @@ import { SearchBarButton } from "@/components/search-bar";
 import DogSelector, { Dog } from "@/components/DogSelector";
 import VetStats from "@/components/event-stats-cards/VetStats";
 import HealthStats from "@/components/event-stats-cards/HealthStats";
-import { FloatingActionButtonVet } from "@/components/floating-action-button-vet";
+import { FloatingActionButtonVet } from "@/components/fabs/floating-action-button-vet";
 
 // Import the events array and the new LandingEventGrid for the bottom cards.
 import { events } from "@/constants/navigation";
