+++ components/action-menu.tsx
@@ -5,8 +5,9 @@ import { motion, AnimatePresence } from "framer-motion";
 import { useRouter } from "next/navigation";
 import type { EventCard } from "@/types";
 import { Modal } from "./ui/modal";
-import { AddEventForm } from "@/components/add-event-form";
-import { VetEventsForm } from "@/components/VetEventsForm";
+import { AddEventForm } from "@/components/add-event-forms/add-event-form";
+import { VetEventsForm } from "@/components/add-event-forms/VetEventsForm";
+import { DietEventsForm } from "@/components/add-event-forms/DietEventsForm";
 
 interface ActionMenuProps {
   isOpen: boolean;
@@ -16,20 +17,28 @@ interface ActionMenuProps {
   onRefresh?: () => void; // New prop to trigger refresh
 }
 
-// Updated: include "Health" and "Poop Journal" as a vet-related event type so they use VetEventsForm
+// Define event type sets.
 const vetEventTypes = new Set([
   "Vet Appointment",
   "Vaccination Appointment",
   "Weight Change",
   "Health",
-  "Poop Journal"
+  "Poop Journal",
 ]);
+
+const dietEventTypes = new Set(["Diet Exception", "Diet Schedule"]);
+
+// Define excluded event types.
+const excludedEventTypes = new Set(["Vet Hub", "Diet"]);
 
 export function ActionMenu({ isOpen, onClose, events, dogId, onRefresh }: ActionMenuProps) {
   const [selectedEvent, setSelectedEvent] = useState<EventCard | null>(null);
   const router = useRouter();
 
-  // A helper callback that closes the modal and menu immediately after a save
+  // Filter out excluded event types.
+  const filteredEvents = events.filter(event => !excludedEventTypes.has(event.title));
+
+  // A helper callback that closes the modal and menu immediately after a save.
   const handleSuccess = () => {
     setSelectedEvent(null);
     onClose();
@@ -56,7 +65,7 @@ export function ActionMenu({ isOpen, onClose, events, dogId, onRefresh }: Action
               exit={{ opacity: 0, y: 20 }}
               className="fixed right-0 bottom-40 flex flex-col-reverse items-end gap-4 z-50 pr-4"
             >
-              {events.map((event, index) => (
+              {filteredEvents.map((event, index) => (
                 <motion.button
                   key={event.title}
                   initial={{ scale: 0, opacity: 0 }}
@@ -68,7 +77,7 @@ export function ActionMenu({ isOpen, onClose, events, dogId, onRefresh }: Action
                   exit={{
                     scale: 0,
                     opacity: 0,
-                    transition: { delay: (events.length - 1 - index) * 0.05 },
+                    transition: { delay: (filteredEvents.length - 1 - index) * 0.05 },
                   }}
                   className="flex items-center justify-end gap-3 w-full"
                   onClick={() => {
@@ -96,9 +105,22 @@ export function ActionMenu({ isOpen, onClose, events, dogId, onRefresh }: Action
         onClose={() => setSelectedEvent(null)}
         title={selectedEvent ? selectedEvent.title : ""}
       >
-        {selectedEvent && vetEventTypes.has(selectedEvent.title) ? (
+        {selectedEvent && dietEventTypes.has(selectedEvent.title) ? (
+          <DietEventsForm
+            eventType={selectedEvent.title as "Diet Exception" | "Diet Schedule"}
+            dogId={dogId!}
+            onSuccess={handleSuccess}
+          />
+        ) : selectedEvent && vetEventTypes.has(selectedEvent.title) ? (
           <VetEventsForm
-            eventType={selectedEvent.title as "Vet Appointment" | "Vaccination Appointment" | "Weight Change" | "Health" | "Poop Journal"}
+            eventType={
+              selectedEvent.title as
+                | "Vet Appointment"
+                | "Vaccination Appointment"
+                | "Weight Change"
+                | "Health"
+                | "Poop Journal"
+            }
             dogId={dogId!}
             onSuccess={handleSuccess}
           />
